Move reporting data source to reporting project
