+++ CDP4Reporting.Tests/DataSource/DataSourceParameterTestFixture.cs
@@ -23,7 +23,7 @@
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace CDP4Composition.Tests.Reporting
+namespace CDP4Reporting.Tests.DataSource
 {
     using System;
     using System.Collections.Concurrent;
@@ -36,7 +36,7 @@ namespace CDP4Composition.Tests.Reporting
     using CDP4Common.SiteDirectoryData;
     using CDP4Common.Types;
 
-    using CDP4Composition.Reporting;
+    using CDP4Reporting.DataSource;
 
     using NUnit.Framework;
