+++ CDP4Reporting/DataSource/IReportingDataSource.cs
@@ -1,5 +1,5 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="DefinedThingShortNameAttributeTestFixture.cs" company="RHEA System S.A.">
+// <copyright file="ICDP4ReportingDataSource.cs" company="RHEA System S.A.">
 //    Copyright (c) 2015-2020 RHEA System S.A.
 //
 //    Author: Sam Gerené, Alex Vorobiev, Alexander van Delft, Cozmin Velciu, Adrian Chivu
@@ -23,20 +23,31 @@
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace CDP4Composition.Tests.Reporting
+namespace CDP4Reporting.DataSource
 {
-    using CDP4Composition.Reporting;
+    using CDP4Common.EngineeringModelData;
+    using CDP4Common.SiteDirectoryData;
 
-    using NUnit.Framework;
-
-    [TestFixture]
-    public class DefinedThingShortNameAttributeTestFixture
+    /// <summary>
+    /// The interface used for creating a reporting data source.
+    /// </summary>
+    /// <typeparam name="T">
+    /// The <see cref="ReportingDataSourceRow"/> representing the data source rows.
+    /// </typeparam>
+    public interface IReportingDataSource<T> where T : ReportingDataSourceRow, new()
     {
-        [Test]
-        public void TestParameterTypeShortNameAttribute()
-        {
-            var attribute = new DefinedThingShortNameAttribute("shortName");
-            Assert.AreEqual(attribute.ShortName, "shortName");
-        }
+        /// <summary>
+        /// Creates a new data source instance.
+        /// </summary>
+        /// <param name="option">
+        /// The <see cref="Option"/> for which the data source is built.
+        /// </param>
+        /// <param name="domainOfExpertise">
+        /// The <see cref="DomainOfExpertise"/> for which the data source is built.
+        /// </param>
+        /// <returns>
+        /// A new <see cref="ReportingDataSourceClass{T}"/> instance.
+        /// </returns>
+        ReportingDataSourceClass<T> CreateDataSource(Option option, DomainOfExpertise domainOfExpertise);
     }
 }
