+++ CDP4Reporting/DataSource/ReportingDataSourceColumn.cs
@@ -23,7 +23,7 @@
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace CDP4Composition.Reporting
+namespace CDP4Reporting.DataSource
 {
     using System;
     using System.Linq;
