+++ CDP4Reporting/DataSource/ReportingDataSourceNode.cs
@@ -23,7 +23,7 @@
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace CDP4Composition.Reporting
+namespace CDP4Reporting.DataSource
 {
     using CDP4Common.EngineeringModelData;
     using CDP4Common.SiteDirectoryData;
