+++ CDP4Reporting/DataSource/ReportingDataSourceParameter.cs
@@ -23,7 +23,7 @@
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace CDP4Composition.Reporting
+namespace CDP4Reporting.DataSource
 {
     using System.Collections.Generic;
     using System.Linq;
